+++ monai/handlers/mean_dice.py
@@ -14,7 +14,7 @@ from typing import Callable, Union, Optional, Sequence
 from ignite.exceptions import NotComputableError
 from ignite.metrics import Metric
 from ignite.metrics.metric import sync_all_reduce, reinit__is_reduced
-from monai.utils.compute_meandice import compute_meandice
+from monai.metrics.compute_meandice import compute_meandice
 
 __all__ = [
     'MeanDice'
