+++ monai/transforms/transforms.py
@@ -9,10 +9,79 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import numpy as np
-import monai
 
-export = monai.utils.export("monai.data.transforms")
+import torch
+import numpy as np
+
+import monai
+from monai.transforms.utils import rescale_array
+from monai.data.utils import get_random_patch, get_valid_patch_size
+
+export = monai.utils.export("monai.transforms")
+
+
+@export
+class AddChannel:
+    """
+    Adds a 1-length channel dimension to the input image.
+    """
+
+    def __call__(self, img):
+        return img[None]
+
+
+@export
+class Transpose:
+    """
+    Transposes the input image based on the given `indices` dimension ordering.
+    """
+
+    def __init__(self, indices):
+        self.indices = indices
+
+    def __call__(self, img):
+        return img.transpose(self.indices)
+
+
+@export
+class Rescale:
+    """
+    Rescales the input image to the given value range.
+    """
+
+    def __init__(self, minv=0.0, maxv=1.0, dtype=np.float32):
+        self.minv = minv
+        self.maxv = maxv
+        self.dtype = dtype
+
+    def __call__(self, img):
+        return rescale_array(img, self.minv, self.maxv, self.dtype)
+
+
+@export
+class ToTensor:
+    """
+    Converts the input image to a tensor without applying any other transformations.
+    """
+
+    def __call__(self, img):
+        return torch.from_numpy(img)
+
+
+@export
+class UniformRandomPatch:
+    """
+    Selects a patch of the given size chosen at a uniformly random position in the image.
+    """
+
+    def __init__(self, patch_size):
+        self.patch_size = (None,) + tuple(patch_size)
+
+    def __call__(self, img):
+        patch_size = get_valid_patch_size(img.shape, self.patch_size)
+        slices = get_random_patch(img.shape, patch_size)
+
+        return img[slices]
 
 
 @export
