+++ monai/utils/__init__.py
@@ -11,4 +11,4 @@
 
 # have to explicitly bring these in here to resolve circular import issues
 from .aliases import alias, resolve_name
-from .moduleutils import export
+from .module import export
