+++ tests/test_dice_loss.py
@@ -14,7 +14,7 @@ import unittest
 import torch
 from parameterized import parameterized
 
-from monai.networks.losses.dice import DiceLoss
+from monai.losses import DiceLoss
 
 TEST_CASE_1 = [  # shape: (1, 1, 2, 2), (1, 1, 2, 2)
     {
